feat: Add validation for guild before create new guild
