+++ src/repositories/guildRepository.ts
@@ -2,4 +2,5 @@ import { Guild, Prisma } from '@prisma/client'
 
 export interface GuildRepository {
   create(guild: Prisma.GuildUncheckedCreateInput): Promise<Guild>
+  findByGuildId(guildId: string): Promise<Guild | null>
 }
